Compiles with debug flag

+++ source/gdul/job_handler/job_handler.cpp
@@ -68,7 +68,7 @@ void job_handler::dump_job_time_sets(const char* location)
 	m_impl->dump_job_time_sets(location);
 }
 #endif
-job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, std::size_t variationId, [[maybe_unused]] const char* dbgName, std::size_t physicalId, [[maybe_unused]] const char* dbgFile, [[maybe_unused]] std::size_t line)
+job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, std::size_t variationId, [[maybe_unused]] const char* dbgName, std::size_t physicalId, [[maybe_unused]] const char* dbgFile, [[maybe_unused]] std::uint32_t line)
 {
 #if defined (GDUL_JOB_DEBUG)
 	return m_impl->make_job_internal(std::move(workUnit), target, physicalId, variationId, dbgName, dbgFile, line);
@@ -76,15 +76,15 @@ job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target
 	return m_impl->make_job_internal(std::move(workUnit), target, physicalId, variationId);
 #endif
 }
-job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, std::size_t variationId, std::size_t physicalId, const char* dbgFile, std::size_t line)
+job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, std::size_t variationId, std::size_t physicalId, const char* dbgFile, std::uint32_t line)
 {
 	return _redirect_make_job(std::move(workUnit), target, variationId, "", physicalId, dbgFile, line);
 }
-job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, const char* dbgName, std::size_t physicalId, const char* dbgFile, std::size_t line)
+job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, const char* dbgName, std::size_t physicalId, const char* dbgFile, std::uint32_t line)
 {
 	return _redirect_make_job(std::move(workUnit), target, 0, dbgName, physicalId, dbgFile, line);
 }
-job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, std::size_t physicalId, const char* dbgFile, std::size_t line)
+job job_handler::_redirect_make_job(delegate<void()> workUnit, job_queue* target, std::size_t physicalId, const char* dbgFile, std::uint32_t line)
 {
 	return _redirect_make_job(std::move(workUnit), target, 0, "", physicalId, dbgFile, line);
 }
